Use Flex to Match Desired Outcome Image
Create divs in index.html to divide into two sections:
title and info boxes. Set body's text alignment to
center to make all elements contained in it
are horizontally centered. Set the bottom margin
of title to 32px to ensure it is not too close to
the info boxes. Make 'plants', the container for
the info, a flex display. Horizontally justify
its content to be centered. Create 52px gap between
all of its elements. Make max width of each info
box 200px wide.